STASH broken attempt to use options in messenger
[skip ci]

The original code called *all* modifiers including pairs that override
each other, such as `min_theta` and `max_cos_theta`, either with a
default value or a value explicitly passed by the user.

We should only call the method if the user specified a value for it
explicitly, and we tried using std::optional, but the G4 messenger
machinery does not understand these.

We need to get on with more urgent stuff right now, so this is
half-done.